+++ src/main/java/com/infoshareacademy/App.java
@@ -1,5 +1,7 @@
 package com.infoshareacademy;
 
+import com.google.gson.Gson;
+
 import java.io.IOException;
 
 import static com.infoshareacademy.menu.Menu.menu;
@@ -12,5 +14,7 @@ public class App
 {
     public static void main( String[] args ) throws IOException {
         menu();
+        Gson gson = new Gson();
+
     }
 }
