+++ src/main/java/com/infoshareacademy/domain/Order.java
@@ -1,4 +1,4 @@
-package com.infoshareacademy.objects;
+package com.infoshareacademy.domain;
 
 public class Order {
     int orderId;
